refactor(account-manager): and fix some lints
