+++ node/standalone/src/rpc.rs
@@ -55,7 +55,7 @@ where
     io.extend_with(TransactionPaymentApi::to_delegate(TransactionPayment::new(
         client.clone(),
     )));
-    io.extend_with(ContractsApi::to_delegate(Contracts::new(client.clone())));
+    io.extend_with(ContractsApi::to_delegate(Contracts::new(client)));
 
     io
 }
