+++ runtime/standalone/src/lib.rs
@@ -16,10 +16,7 @@ use sp_consensus_aura::sr25519::AuthorityId as AuraId;
 use sp_core::{crypto::KeyTypeId, OpaqueMetadata};
 use sp_runtime::{
     create_runtime_str, generic, impl_opaque_keys,
-    traits::{
-        AccountIdLookup, BlakeTwo256, Block as BlockT, IdentifyAccount, NumberFor, Saturating,
-        Verify,
-    },
+    traits::{AccountIdLookup, BlakeTwo256, Block as BlockT, IdentifyAccount, NumberFor, Verify},
     transaction_validity::{TransactionSource, TransactionValidity},
     ApplyExtrinsicResult, MultiSignature,
 };
@@ -29,7 +26,6 @@ use sp_version::NativeVersion;
 use sp_version::RuntimeVersion;
 
 // A few exports that help ease life for downstream crates.
-use frame_support::traits::IsSubType;
 pub use frame_support::{
     construct_runtime, parameter_types,
     traits::{
@@ -44,7 +40,6 @@ pub use frame_support::{
 };
 pub use pallet_balances::Call as BalancesCall;
 pub use pallet_timestamp::Call as TimestampCall;
-use pallet_transaction_payment::CurrencyAdapter;
 #[cfg(any(feature = "std", test))]
 pub use sp_runtime::BuildStorage;
 pub use sp_runtime::{Perbill, Permill};
@@ -269,97 +264,8 @@ impl pallet_balances::Config for Runtime {
     type WeightInfo = pallet_balances::weights::SubstrateWeight<Runtime>;
 }
 
-pub struct AccountManagerCurrencyAdapter<C, OU>(sp_std::marker::PhantomData<(C, OU)>);
+pallet_account_manager::setup_currency_adapter!();
 
-impl<T, C, OU> pallet_transaction_payment::OnChargeTransaction<T>
-    for AccountManagerCurrencyAdapter<C, OU>
-where
-    T: pallet_transaction_payment::Config + pallet_account_manager::Config<Currency = C>,
-    T::TransactionByteFee: frame_support::pallet_prelude::Get<<C as frame_support::traits::Currency<<T as frame_system::Config>::AccountId>>::Balance>,
-    <T as frame_system::Config>::Call: IsSubType<pallet_3vm_contracts::Call<Runtime>>,
-    C: frame_support::traits::Currency<<T as frame_system::Config>::AccountId>,
-    C::PositiveImbalance: Imbalance<
-        <C as frame_support::traits::Currency<<T as frame_system::Config>::AccountId>>::Balance,
-        Opposite = C::NegativeImbalance,
-    >,
-    C::NegativeImbalance: Imbalance<
-        <C as frame_support::traits::Currency<<T as frame_system::Config>::AccountId>>::Balance,
-        Opposite = C::PositiveImbalance,
-    >,
-    OU: OnUnbalanced<<C as frame_support::traits::Currency<<T as frame_system::Config>::AccountId>>::NegativeImbalance>,
-{
-    type Balance =
-        <C as frame_support::traits::Currency<<T as frame_system::Config>::AccountId>>::Balance;
-    type LiquidityInfo = (Option<<T as frame_system::Config>::AccountId>, Option<pallet_account_manager::transaction::NegativeImbalanceOf<T, C>>); // TODO: modify this to be (Account, Imbalance)
-
-    fn withdraw_fee(
-        who: &T::AccountId,
-        call: &<T as frame_system::Config>::Call,
-        info: &sp_runtime::traits::DispatchInfoOf<<T as frame_system::Config>::Call>,
-        fee: Self::Balance,
-        tip: Self::Balance,
-    ) -> Result<Self::LiquidityInfo, frame_support::pallet_prelude::TransactionValidityError> {
-        let call: &<T as frame_system::Config>::Call = call;
-
-        let result = <CurrencyAdapter<C, OU> as pallet_transaction_payment::OnChargeTransaction<T>>::withdraw_fee(
-            who, call, info, fee, tip,
-        );
-
-        if let Some(pallet_3vm_contracts::Call::call { dest, .. }) = call.is_sub_type() {
-            if let Some(author) = Contracts::get_author(dest) {
-                result.map(|info| {
-                    let opaque_author = T::AccountId::decode(&mut author.as_ref()).unwrap(); // TODO: remove
-                    (Some(opaque_author), info)
-                })
-            } else {
-                result.map(|info| (None, info))
-            }
-        } else {
-            result.map(|info| (None, info))
-        }
-    }
-
-    fn correct_and_deposit_fee(
-        who: &T::AccountId,
-        _dispatch_info: &sp_runtime::traits::DispatchInfoOf<<T as frame_system::Config>::Call>,
-        _post_info: &sp_runtime::traits::PostDispatchInfoOf<<T as frame_system::Config>::Call>,
-        corrected_fee: Self::Balance,
-        tip: Self::Balance,
-        already_withdrawn: Self::LiquidityInfo,
-    ) -> Result<(), frame_support::pallet_prelude::TransactionValidityError> {
-        if let (beneficiary, Some(paid)) = already_withdrawn {
-            // Calculate how much refund we should return
-            let refund_amount = paid.peek().saturating_sub(corrected_fee);
-            // refund to the the account that paid the fees. If this fails, the
-            // account might have dropped below the existential balance. In
-            // that case we don't refund anything.
-            let refund_imbalance = C::deposit_into_existing(who, refund_amount)
-                .unwrap_or_else(|_| C::PositiveImbalance::zero());
-            // merge the imbalance caused by paying the fees and refunding parts of it again.
-            let adjusted_paid = paid.offset(refund_imbalance).same().map_err(|_| {
-                frame_support::pallet_prelude::TransactionValidityError::Invalid(
-                    frame_support::pallet_prelude::InvalidTransaction::Payment,
-                )
-            })?;
-
-            // Call someone else to handle the imbalance (fee and tip separately)
-            let (tip, fee) = adjusted_paid.split(tip);
-
-            // Since there is a beneficiary, account-manager handles it
-            if let Some(beneficiary) = beneficiary {
-                let fee_with_tip = Some(fee).into_iter().chain(Some(tip));
-                let as_beneficiaries = fee_with_tip
-                    .map(|i| {
-                        pallet_account_manager::transaction::ImbalanceBeneficiary::<T, C>(Some(beneficiary.clone()), i)
-                    });
-                <pallet_account_manager::transaction::ImbalanceToBeneficiary<T> as OnUnbalanced<pallet_account_manager::transaction::ImbalanceBeneficiary<T, C>>>::on_unbalanceds(as_beneficiaries);
-            } else {
-                <OU as OnUnbalanced<<C as frame_support::traits::Currency<<T as frame_system::Config>::AccountId>>::NegativeImbalance>>::on_unbalanceds(Some(fee).into_iter().chain(Some(tip)));
-            }
-        }
-        Ok(())
-    }
-}
 parameter_types! {
     pub const TransactionByteFee: Balance = 1;
 }
